% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_api.R
\name{api_get}
\alias{api_get}
\title{Get meta data details for specified Google API}
\usage{
api_get(api, version)
}
\arguments{
\item{api}{The API to fetch}

\item{version}{The API version to fetch}
}
\value{
Details of the API
}
\description{
Download the discovery document for an API
}
\seealso{
\url{https://developers.google.com/discovery/v1/getting_started}
}
\concept{Google Discovery API functions}
