% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_api.R
\name{api_list}
\alias{api_list}
\title{Get a list of Google API libraries}
\usage{
api_list(name = NULL, preferred = NULL)
}
\arguments{
\item{name}{Only include APIs with the given name. *Case-sensitive. All are lowercase currently.}

\item{preferred}{Return only the preferred version of an API.  FALSE by default.}
}
\value{
List of Google APIs and their resources
}
\description{
Does not require authentication
}
\seealso{
\url{https://developers.google.com/discovery/v1/reference/apis/list}

Other API Discovery Service: 
\code{\link{package_make}()}
}
\concept{API Discovery Service}
