% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_api.R
\name{get_google_api}
\alias{get_google_api}
\title{Get meta data details for specified Google API}
\usage{
get_google_api(api, version)
}
\arguments{
\item{api}{The API to fetch}

\item{version}{The API version to fetch}

\item{a_url}{Supply your own discovery URL, for private APIs only}
}
\value{
Details of the API
}
\description{
Download the discovery document for an API
}
\seealso{
\url{https://developers.google.com/discovery/v1/getting_started}

Other Google Discovery API functions: 
\code{\link{list_google_apis}()},
\code{\link{make_google_package}()}
}
\concept{Google Discovery API functions}
