% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery_api.R
\name{list_google_apis}
\alias{list_google_apis}
\title{Get a list of Google API libraries}
\usage{
list_google_apis(name = NULL, preferred = NULL)
}
\arguments{
\item{name}{Only include APIs with the given name. *Case-sensitive. All are lowercase currently.}

\item{preferred}{Return only the preferred version of an API.  "false" by default.}
}
\value{
List of Google APIs and their resources
}
\description{
Does not require authentication
}
\seealso{
\url{https://developers.google.com/discovery/v1/reference/apis/list}

Other Google Discovery API functions: 
\code{\link{get_google_api}()},
\code{\link{make_google_package}()}
}
\concept{Google Discovery API functions}
