% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{make_google_package}
\alias{make_google_package}
\title{Create a Google API package}
\usage{
make_google_package(api_id, output_dir, package_name)
}
\arguments{
\item{output_dir}{Directory path to write the package to.}

\item{package_name}{Name of the package to create. NOTE: {output_dir}/{package_name} cannot be pre-existing.}
}
\value{
TRUE if successful, side effect will write package directory {output_dir}/{package_name}.
}
\description{
This function will generate a package of Google API functions.
}
\seealso{
Other Google Discovery API functions: 
\code{\link{get_google_api}()},
\code{\link{list_google_apis}()}
}
\concept{Google Discovery API functions}
