% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{package_make}
\alias{package_make}
\title{Create a Google API package}
\usage{
package_make(api_id, output_dir, package_name)
}
\arguments{
\item{api_id}{The api to fetch. Run \code{\link{api_list}} for options.}

\item{output_dir}{Directory path to write the package to.}

\item{package_name}{Name of the package to create. NOTE: output_dir/package_name cannot be pre-existing.}
}
\value{
TRUE if successful, side effect will write package directory output_dir/package_name.
}
\description{
This function will generate a package of Google API functions.
}
\note{
You will still need to run devtools::document(your_package_directory) and install your package for testing.
}
\seealso{
Other Google Discovery API functions: 
\code{\link{api_get}()},
\code{\link{api_list}()}
}
\concept{Google Discovery API functions}
